feat(api): add media.history, media.lookup, and media.control endpoints
Implements planned media API improvements for enhanced game discovery
and active media control in the Zaparoo App.

New API methods:

- media.history: cursor-paginated play history from UserDB with
  configurable limit (default 25, max 100). Exposes system, media name,
  path, launcher, timestamps, and play time.

- media.lookup: single-title resolution using the full 6-strategy fuzzy
  matching pipeline. Returns best match with confidence score, system
  metadata, and ready-to-use ZapScript. Returns null match (not error)
  when below confidence threshold.

- media.control: dispatches control actions (save_state, load_state,
  toggle_menu, etc.) to the active media's launcher. Supports both
  native Go functions and ZapScript-based controls. Validates all
  commands before executing any.

Launcher controls system:

- Launchers declare capabilities via a Controls map (action -> func or
  script). Nil map = no capabilities, so existing launchers need zero
  changes.
- KeyboardControls helper for the common keyboard-shortcut case.
- Custom TOML launchers support a [commands] section mapping actions to
  ZapScript strings (e.g., save_state = "keyboard:{f2}").
- media.active response now includes launcherControls listing available
  actions for the active launcher.
- RunControlScript executes ZapScript in a restricted context that
  blocks media-launching and playlist commands.

Supporting changes:

- Extract title resolution pipeline from cmdTitle into reusable
  titles.ResolveTitle function with ResolveParams/ResolveResult types
  and ErrNoMatch/ErrLowConfidence sentinel errors.
- Centralize system ID resolution into resolveSystem/resolveSystems
  helpers with fuzzy matching support via new FuzzySystem param.
- Centralize ZapScript title construction in BuildTitleZapScript.
- Enrich media.stopped notification with system, media, launcher, and
  elapsed time (was nil payload).
- Move playlist-clear-on-media-change from RunCommand to service layer
  so control scripts can safely reuse RunCommand.

Bug fixes:

- Fix TOCTOU race in indexing status: startIfNotRunning atomically
  checks and sets under a single lock.
- Fix pagination sentinel overflow: int -> int64 for history lastID
  throughout the database interface chain.
- Clamp elapsed time to zero in media.stopped to handle NTP clock
  corrections on embedded devices.
- Drop synthetic ScanTime from control script tokens so pipeline guards
  reject them if they ever leak into the scan queue.
- Fix shared GlobalLauncherCache test concurrency with explicit mutex.

+++ pkg/api/methods/playtime_test.go
@@ -77,7 +77,7 @@ func TestHandlePlaytime_ResetStateWithDailyFields(t *testing.T) {
 	t.Parallel()
 
 	mockDB := testhelpers.NewMockUserDBI()
-	mockDB.On("GetMediaHistory", 0, 100).Return([]database.MediaHistoryEntry{
+	mockDB.On("GetMediaHistory", int64(0), 100).Return([]database.MediaHistoryEntry{
 		{
 			DBID:      1,
 			StartTime: time.Date(2025, 1, 15, 10, 0, 0, 0, time.UTC),
@@ -176,7 +176,7 @@ func TestHandlePlaytime_CooldownStateWithDailyFields(t *testing.T) {
 	t.Parallel()
 
 	mockDB := testhelpers.NewMockUserDBI()
-	mockDB.On("GetMediaHistory", 0, 100).Return([]database.MediaHistoryEntry{
+	mockDB.On("GetMediaHistory", int64(0), 100).Return([]database.MediaHistoryEntry{
 		{
 			DBID:      1,
 			StartTime: time.Date(2025, 1, 15, 10, 0, 0, 0, time.UTC),
